Removed unnecessary casting (4)

+++ SF3Lib/Models/Structs/MPD/Main/Gradient.cs
@@ -47,63 +47,63 @@ namespace SF3.Models.Structs.MPD.Main {
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_topPositionAddr), displayName: "TopPosition", displayOrder: 0.1f, displayFormat: "X2")]
         public ushort TopPositionRaw {
-            get => (ushort) Data.GetUInt16(_topPositionAddr);
+            get => Data.GetUInt16(_topPositionAddr);
             set => Data.SetUInt16(_topPositionAddr, value);
         }
 
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_bottomPositionAddr), displayName: "BottomPosition", displayOrder: 1, displayFormat: "X2")]
         public ushort BottomPositionRaw {
-            get => (ushort) Data.GetUInt16(_bottomPositionAddr);
+            get => Data.GetUInt16(_bottomPositionAddr);
             set => Data.SetUInt16(_bottomPositionAddr, value);
         }
 
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_topRAddr), displayOrder: 2, displayFormat: "X2")]
         public ushort TopR {
-            get => (ushort) Data.GetUInt16(_topRAddr);
+            get => Data.GetUInt16(_topRAddr);
             set => Data.SetUInt16(_topRAddr, value);
         }
 
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_topGAddr), displayOrder: 3, displayFormat: "X2")]
         public ushort TopG {
-            get => (ushort) Data.GetUInt16(_topGAddr);
+            get => Data.GetUInt16(_topGAddr);
             set => Data.SetUInt16(_topGAddr, value);
         }
 
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_topBAddr), displayOrder: 4, displayFormat: "X2")]
         public ushort TopB {
-            get => (ushort) Data.GetUInt16(_topBAddr);
+            get => Data.GetUInt16(_topBAddr);
             set => Data.SetUInt16(_topBAddr, value);
         }
 
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_bottomRAddr), displayOrder: 5, displayFormat: "X2")]
         public ushort BottomR {
-            get => (ushort) Data.GetUInt16(_bottomRAddr);
+            get => Data.GetUInt16(_bottomRAddr);
             set => Data.SetUInt16(_bottomRAddr, value);
         }
 
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_bottomGAddr), displayOrder: 6, displayFormat: "X2")]
         public ushort BottomG {
-            get => (ushort) Data.GetUInt16(_bottomGAddr);
+            get => Data.GetUInt16(_bottomGAddr);
             set => Data.SetUInt16(_bottomGAddr, value);
         }
 
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_bottomBAddr), displayOrder: 7, displayFormat: "X2")]
         public ushort BottomB {
-            get => (ushort) Data.GetUInt16(_bottomBAddr);
+            get => Data.GetUInt16(_bottomBAddr);
             set => Data.SetUInt16(_bottomBAddr, value);
         }
 
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_partsAffectedBitsAddr), displayOrder: 8, displayFormat: "X1")]
         public ushort PartsAffectedBits {
-            get => (ushort) Data.GetUInt16(_partsAffectedBitsAddr);
+            get => Data.GetUInt16(_partsAffectedBitsAddr);
             set => Data.SetUInt16(_partsAffectedBitsAddr, value);
         }
 
@@ -128,21 +128,21 @@ namespace SF3.Models.Structs.MPD.Main {
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_groundIntensityAddr), displayName: "GroundItensity", displayOrder: 9, displayFormat: "X2")]
         public ushort GroundIntensityRaw {
-            get => (ushort) Data.GetUInt16(_groundIntensityAddr);
+            get => Data.GetUInt16(_groundIntensityAddr);
             set => Data.SetUInt16(_groundIntensityAddr, value);
         }
 
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_skyIntensityAddr), displayName: "SkyIntensity", displayOrder: 10, displayFormat: "X2")]
         public ushort SkyIntensityRaw {
-            get => (ushort) Data.GetUInt16(_skyIntensityAddr);
+            get => Data.GetUInt16(_skyIntensityAddr);
             set => Data.SetUInt16(_skyIntensityAddr, value);
         }
 
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_modelsAndSurfaceIntensityAddr), displayName: "ModelsAndSurfaceIntensity", displayOrder: 11, displayFormat: "X2")]
         public ushort ModelsAndSurfaceIntensityRaw {
-            get => (ushort) Data.GetUInt16(_modelsAndSurfaceIntensityAddr);
+            get => Data.GetUInt16(_modelsAndSurfaceIntensityAddr);
             set => Data.SetUInt16(_modelsAndSurfaceIntensityAddr, value);
         }
